requirements.txt
praw==7.7.1
prawcore==2.4.0
pandas==2.1.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.2
mysql-connector-python
plotly
streamlit